docs: add doxygen comments

+++ src/visual/ellipse.cpp
@@ -11,8 +11,8 @@ Ellipse::Ellipse(const cv::Point2f center, const cv::Mat& matrix) {
 }
 
 cv::Size2f Ellipse::Axes() {
-  const cv::Size2f axes(static_cast<double>(2.0L * sqrt(eigen_values_.at<double>(0, 0) * CHISQ_95_2)),
-                        static_cast<double>(2.0L * sqrt(eigen_values_.at<double>(0, 0) * CHISQ_95_2)));
+  const cv::Size2f axes(static_cast<float>(2.0L * sqrt(eigen_values_.at<double>(0, 0) * CHISQ_95_2)),
+                        static_cast<float>(2.0L * sqrt(eigen_values_.at<double>(0, 0) * CHISQ_95_2)));
 
   return axes;
 }
